# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ChainInfoResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"chains": {
						"items": {
							"properties": {
								"rpcUrl": {
									"type": "string"
								},
								"explorerUrl": {
									"type": "string"
								},
								"currency": {
									"type": "string"
								},
								"testnet": {
									"type": "boolean"
								},
								"name": {
									"type": "string"
								},
								"chainId": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"rpcUrl",
								"explorerUrl",
								"currency",
								"testnet",
								"name",
								"chainId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"chains"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NativeCurrency": {
				"properties": {
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"symbol",
					"decimals"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SupportedChainsResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"chains": {
						"items": {
							"properties": {
								"nativeCurrency": {
									"$ref": "#/components/schemas/NativeCurrency"
								},
								"rpcUrl": {
									"type": "string"
								},
								"explorerUrl": {
									"type": "string"
								},
								"currency": {
									"type": "string"
								},
								"testnet": {
									"type": "boolean"
								},
								"name": {
									"type": "string"
								},
								"chainId": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"nativeCurrency",
								"rpcUrl",
								"explorerUrl",
								"currency",
								"testnet",
								"name",
								"chainId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"chains"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NativeBalanceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"balance": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NativeBalanceRequest": {
				"properties": {
					"walletAddress": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"walletAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenBalance": {
				"properties": {
					"symbol": {
						"type": "string"
					},
					"balance": {
						"type": "string"
					},
					"formattedBalance": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"tokenAddress": {
						"type": "string"
					},
					"isNative": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"symbol",
					"balance",
					"formattedBalance",
					"decimals",
					"tokenAddress",
					"isNative"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenBalanceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"tokenBalance": {
						"$ref": "#/components/schemas/TokenBalance"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BalanceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"walletAddress": {
						"type": "string"
					},
					"nativeBalance": {
						"type": "string"
					},
					"tokenBalances": {
						"items": {
							"$ref": "#/components/schemas/TokenBalance"
						},
						"type": "array"
					},
					"totalValueUSD": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"walletAddress",
					"nativeBalance",
					"tokenBalances",
					"chainId",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BalanceRequest": {
				"properties": {
					"walletAddress": {
						"type": "string"
					},
					"tokenAddress": {
						"type": "string"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"walletAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MultipleTokensResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"walletAddress": {
						"type": "string"
					},
					"tokenBalances": {
						"items": {
							"$ref": "#/components/schemas/TokenBalance"
						},
						"type": "array"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"walletAddress",
					"tokenBalances",
					"chainId",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MultipleTokensRequest": {
				"properties": {
					"walletAddress": {
						"type": "string"
					},
					"tokenAddresses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"walletAddress",
					"tokenAddresses"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"app": {
				"type": "apiKey",
				"name": "X-App-ID",
				"in": "header",
				"description": "Application ID for API access"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT token for authentication"
			}
		}
	},
	"info": {
		"title": "crefy-connect-backend",
		"version": "1.0.0",
		"description": "Wallet authentication API with Express and TypeScript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jimleston Osoi"
		}
	},
	"paths": {
...
		"/signing/chains": {
			"get": {
				"operationId": "GetSupportedChains",
				"responses": {
					"200": {
						"description": "Supported chains retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChainInfoResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"chains": [
												{
													"chainId": 1,
													"name": "Ethereum",
													"testnet": false,
													"currency": "ETH",
													"explorerUrl": "https://etherscan.io",
													"rpcUrl": "https://eth.llamarpc.com"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all supported chains",
				"tags": [
					"Signing Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
...
		"/balance/chains": {
			"get": {
				"operationId": "GetSupportedChains",
				"responses": {
					"200": {
						"description": "Supported chains retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupportedChainsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"chains": [
												{
													"chainId": 8453,
													"name": "Base",
													"testnet": false,
													"currency": "ETH",
													"explorerUrl": "https://basescan.org",
													"rpcUrl": "https://mainnet.base.org",
													"nativeCurrency": {
														"name": "Ethereum",
														"symbol": "ETH",
														"decimals": 18
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all supported chains for balance checking",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/balance/native": {
			"post": {
				"operationId": "GetNativeBalance",
				"responses": {
					"200": {
						"description": "Native balance retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NativeBalanceResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"balance": "0.125"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get native balance only",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NativeBalanceRequest"
							}
						}
					}
				}
			}
		},
		"/balance/token": {
			"post": {
				"operationId": "GetTokenBalance",
				"responses": {
					"200": {
						"description": "Token balance retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenBalanceResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"tokenBalance": {
												"symbol": "USDC",
												"balance": "1500000000",
												"formattedBalance": "1500.00",
												"decimals": 6,
												"tokenAddress": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
												"isNative": false,
												"name": "USD Coin"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get token balance for a specific ERC20 token",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/balance/all": {
			"post": {
				"operationId": "GetBalances",
				"responses": {
					"200": {
						"description": "Balances retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BalanceResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"walletAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"nativeBalance": "0.125",
											"tokenBalances": [
												{
													"symbol": "ETH",
													"balance": "125000000000000000",
													"formattedBalance": "0.125",
													"decimals": 18,
													"tokenAddress": "0x0000000000000000000000000000000000000000",
													"isNative": true,
													"name": "Ethereum"
												},
												{
													"symbol": "USDC",
													"balance": "1500000000",
													"formattedBalance": "1500.00",
													"decimals": 6,
													"tokenAddress": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
													"isNative": false,
													"name": "USD Coin"
												}
											],
											"chainId": 8453,
											"chainName": "Base"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all balances (native + specified token)",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BalanceRequest"
							}
						}
					}
				}
			}
		},
		"/balance/multiple-tokens": {
			"post": {
				"operationId": "GetMultipleTokenBalances",
				"responses": {
					"200": {
						"description": "Multiple token balances retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MultipleTokensResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"walletAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"tokenBalances": [
												{
													"symbol": "USDC",
													"balance": "1500000000",
													"formattedBalance": "1500.00",
													"decimals": 6,
													"tokenAddress": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
													"isNative": false,
													"name": "USD Coin"
												},
												{
													"symbol": "DAI",
													"balance": "250000000000000000000",
													"formattedBalance": "250.00",
													"decimals": 18,
													"tokenAddress": "0x50c5725949A6F0c72E6C4a641F24049A917DB0Cb",
													"isNative": false,
													"name": "Dai Stablecoin"
												}
											],
											"chainId": 8453,
											"chainName": "Base"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get multiple token balances at once",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MultipleTokensRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}